+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Add `EveOnline::ESI::Alliances`
 * Add `EveOnline::ESI::AllianceCorporations`
+* Add `EveOnline::ESI::Models::AllianceShort`
 
 **v0.14.0**